+++ module-app/build.gradle
@@ -41,6 +41,7 @@ dependencies {
     compileOnly 'org.modelmapper:modelmapper:3.1.0'
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
