+++ module-core/src/main/java/dto/genre/GenreDTO.java
@@ -1,0 +1,9 @@
+package dto.genre;
+
+import lombok.Getter;
+
+@Getter
+public class GenreDTO {
+	private Long genreId;
+	private String genreName;
+}
