--- module-core/src/main/java/dto/internal/GenreDTO.java
@@ -1,9 +1,0 @@
-package dto.internal;
-
-import lombok.Getter;
-
-@Getter
-public class GenreDTO {
-	private Long id;
-	private String name;
-}
