+++ module-core/src/main/java/dto/movie/MovieDTO.java
@@ -1,13 +1,15 @@
-package dto.internal;
+package dto.movie;
 
 import java.time.LocalDate;
 import java.util.Objects;
 
+import dto.genre.GenreDTO;
+import dto.rating.RatingDTO;
 import lombok.Getter;
 
 @Getter
 public class MovieDTO {
-	private Long id;
+	private Long movieId;
 	private String title;
 	private LocalDate openDay;
 	private Integer runningTimeAsMinute;
@@ -21,11 +23,11 @@ public class MovieDTO {
 		if (this == obj) return true; // 같은 객체라면 true
 		if (obj == null || getClass() != obj.getClass()) return false; // 클래스 타입 확인
 		MovieDTO mObj = (MovieDTO) obj; // 안전한 캐스팅
-		return id == mObj.id; // id 비교
+		return movieId == mObj.movieId; // id 비교
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(id); // id를 기준으로 hashCode 생성
+		return Objects.hash(movieId); // id를 기준으로 hashCode 생성
 	}
 }
