+++ module-rds-repo/src/main/java/module/repository/rating/RatingRepository.java
@@ -1,4 +1,4 @@
-package module.repository;
+package module.repository.rating;
 
 import org.springframework.data.jpa.repository.JpaRepository;
