+++ module-service/src/main/java/module/service/genre/GenreService.java
@@ -1,8 +1,8 @@
-package module.service;
+package module.service.genre;
 
 import java.util.List;
 
-import dto.internal.GenreDTO;
+import dto.genre.GenreDTO;
 
 public interface GenreService {
 	List<GenreDTO>  getAllGenre();
